requirements.txt
customtkinter
playwright
beautifulsoup4
httpx
google-generativeai
openai
PyMuPDF
keyring
